%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  argAssertRange,private.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{.argAssertRange}
\alias{.argAssertRange}


\title{A private function}

\description{
  A private function.
}

\usage{.argAssertRange(x, range, ...)}

\arguments{
  \item{x}{A \code{\link[base]{numeric}} \code{\link[base]{vector}} to be validated.}
  \item{range}{A \code{\link[base]{numeric}} \code{\link[base]{vector}} of length two.}
  \item{...}{Not used.}
}

\value{
  Returns \code{x} if valid, otherwise an error is thrown.
}

\author{Henrik Bengtsson (\url{http://www.braju.com/R/})}



\keyword{programming}
\keyword{internal}
