%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
% Do not modify this file since it was automatically generated from:
% 
%  999.package.R
% 
% by the Rdoc compiler part of the R.oo package.
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

 \name{cnagIO-package}
\alias{cnagIO-package}
\alias{cnagIO}
\docType{package}

\title{Package cnagIO}


 \description{
   This package provides methods for reading CFH files generated by the CNAG software.

   This package should be considered to be in an alpha or beta phase.
 }

 \section{Installation and updates}{
   To install this package, see instructions at
   \url{http://www.braju.com/R/}.
 }

 \author{
   Henrik Bengtsson.
 }

 \section{To get started}{
   To get started, see:
   \enumerate{
     \item \code{\link{readCfhUnits}}() - reads a CNAG CFH file.
     \item \code{\link{readCfnUnits}}() - reads a CNAG CFN file.
   }
 }

 \section{License}{
   The releases of this package is licensed under
   LGPL version 2.1 or newer.
 }

 \references{
  [1] The CNAG software, \url{www.genome.umin.jp/}, Oct 2008.
 }
\keyword{package}
